+++ src/main/java/com/monster/repository/UtenteRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.monster.persistence.entity.Utente;
 
 
-public interface UtenteRepository extends JpaRepository<Utente, Integer>{
+public interface UtenteRepository extends JpaRepository<Utente, Long>{
 	
 	Utente findByEmail(String email);
 	//@Query("SELECT u FROM User u WHERE x.age >= ?1")
